Split people/update from people/index
